Use absolute paths to test executables
cet_test's setup of running applications through cet_test_exec can
result in that latter being passed pure target (and hence executable)
names. Unless the executable is present in the PATH, tests will fail.

When building, or using, a CMake target, expand it using a genex to
the full path. When using cet_script, pass the full path to script
as generated.