Remove PATH requirement for scripts running scripts
CheckClassVersion and RunAndCompare use programs that are part of
cetmodules, but execute them using only the program name. This
relies on clients having the path to cetmodules' binary directory
in the PATH. This may not always be the case, and also opens up the
possibility for clashes should conflicting paths be prepended.

Instead, use cetmodules/CMake internal knowledge of program locations
to run them via:

- CheckClassVersion is used after find_package(cetmodules), so use
  cetmodules_bin_dir variable to derive path to checkClassVersion
  program
- RunAndCompare is run as a CMake script, so use known and currently
  fixed install layout of parallel bin/ and Modules/ dirs to locate
  filter-output script "one directory level over"